requirements.txt
anyio==4.9.0
asgiref==3.8.1
attrs==25.3.0
autobahn==24.4.2
Automat==25.4.16
certifi==2025.4.26
cffi==1.17.1
channels==4.1.0
channels_redis==4.2.1
chapa==0.1.2
charset-normalizer==3.4.2
constantly==23.10.4
cryptography==44.0.3
daphne==4.1.2
dj-database-url==2.3.0
Django==5.2
whitenoise==6.9.0
django-allauth==65.7.0
django-cleanup==9.0.0
django-environ==0.12.0
django-htmx==1.23.0
django-telegram-login==0.2.3
h11==0.16.0
httpcore==1.0.9
httpx==0.28.1
hyperlink==21.0.0
idna==3.10
incremental==24.7.2
msgpack==1.1.0
oauthlib==3.2.2
pillow==11.2.1
psycopg2-binary==2.9.10
pyasn1==0.6.1
pyasn1_modules==0.4.2
pycparser==2.22
pyOpenSSL==25.0.0
pytailwindcss==0.2.0
python-dotenv==1.1.0
python-telegram-bot==22.1
redis==6.1.0
requests==2.32.3
requests-oauthlib==2.0.0
service-identity==24.2.0
setuptools==80.3.1
six==1.17.0
sniffio==1.3.1
sqlparse==0.5.3
Twisted==24.11.0
txaio==23.1.1
typing_extensions==4.13.2
tzdata==2025.2
urllib3==2.4.0
zope.interface==7.2
